requirements.txt
# Responsibility: Python dependencies for data processing and exploration notebooks used by "The Stress Galaxy".
pandas
numpy
scikit-learn
umap-learn
matplotlib
seaborn
jupyter
